test(mongo): parametrized tests regarding db version
